# openapi.yml
openapi: 3.1.0
info:
  title: Система бронирования мест в кафе
  version: 0.0.2
servers:
  - url: http://localhost:8000/
    description: Локальный сервер
...
paths:
...
  /cafe/{cafe_id}/time_slots:
    get:
      tags:
        - Временные слоты
      summary: Список временных слотов в кафе
      description: Получение списка доступных для бронирования временных слотов в кафе. Для администраторов и менеджеров - все столы (с возможностью выбора), для пользователей - только активные.
      operationId: get_all_time_slots_cafe__cafe_id__time_slots_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: cafe_id
          in: path
          required: true
          schema:
            type: integer
            description: ID кафе
            title: Cafe Id
          description: ID кафе
        - name: show_all
          in: query
          required: false
          schema:
            type: boolean
            title: Показывать все слоты?
            description: Показывать все временные слоты в кафе или нет. По умолчанию показывает все столы
            default: false
          description: Показывать все временные слоты в кафе или нет. По умолчанию показывает все столы
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlotInfo'
                title: Response Get All Time Slots Cafe  Cafe Id  Time Slots Get
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
    post:
      tags:
        - Временные слоты
      summary: Новый временной слот в кафе
      description: Создает нового временного слота в кафе. Только для администраторов и менеджеров.
      operationId: create_time_slot_cafe__cafe_id__time_slots_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: cafe_id
          in: path
          required: true
          schema:
            type: integer
            description: ID кафе
            title: Cafe Id
          description: ID кафе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeSlotCreate'
              description: Данные нового временного слота
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlotInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
  /cafe/{cafe_id}/time_slots/{slot_id}:
    get:
      tags:
        - Временные слоты
      summary: Информация о временном слоте в кафе по его ID
      description: Получение информации о временном слоте в кафе по его ID. Для администраторов и менеджеров - все столы, для пользователей - только активные.
      operationId: get_time_slot_by_id_cafe__cafe_id__time_slots__slot_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: cafe_id
          in: path
          required: true
          schema:
            type: integer
            description: ID кафе
            title: Cafe Id
          description: ID кафе
        - name: slot_id
          in: path
          required: true
          schema:
            type: integer
            description: ID слота
            title: Slot Id
          description: ID слота
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlotInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
    patch:
      tags:
        - Временные слоты
      summary: Обновление информации о временом слоте в кафе по его ID
      description: Обновление информации о временом слоте в кафе по его ID. Только для администраторов и менеджеров.
      operationId: path_time_slot_by_id_cafe__cafe_id__time_slots__slot_id__patch
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: cafe_id
          in: path
          required: true
          schema:
            type: integer
            description: ID кафе
            title: Cafe Id
          description: ID кафе
        - name: slot_id
          in: path
          required: true
          schema:
            type: integer
            description: ID слота
            title: Slot Id
          description: ID слота
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeSlotUpdate'
              description: Обновленные данные временного слота
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlotInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
...
components:
  schemas:
...
    CafeShortInfo:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        address:
          type: string
          title: Address
        phone:
          type: string
          title: Phone
        description:
          type: string
          title: Description
        photo_id:
          type: string
          format: uuid
          title: Photo Id
      type: object
      required:
        - id
        - name
        - address
        - phone
        - description
        - photo_id
      title: CafeShortInfo
...
    CustomError:
      properties:
        code:
          type: integer
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
        - code
        - message
      title: CustomError
...
    TimeSlotCreate:
      properties:
        cafe_id:
          type: integer
          title: Cafe Id
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          type: string
          format: time
          title: End Time
        description:
          type: string
          title: Description
      type: object
      required:
        - cafe_id
        - start_time
        - end_time
        - description
      title: TimeSlotCreate
    TimeSlotInfo:
      properties:
        id:
          type: integer
          title: Id
        cafe:
          $ref: '#/components/schemas/CafeShortInfo'
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          type: string
          format: time
          title: End Time
        description:
          type: string
          title: Description
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
        - id
        - cafe
        - start_time
        - end_time
        - description
        - is_active
        - created_at
        - updated_at
      title: TimeSlotInfo
...
    TimeSlotUpdate:
      properties:
        cafe_id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Cafe Id
        start_time:
          anyOf:
            - type: string
              format: time
            - type: 'null'
          title: Start Time
        end_time:
          anyOf:
            - type: string
              format: time
            - type: 'null'
          title: End Time
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        is_active:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Is Active
      type: object
      required:
        - cafe_id
        - start_time
        - end_time
        - description
        - is_active
      title: TimeSlotUpdate
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login